Fix panic when formatting timestamp with timezone specifier (#983)
Fixes a panic that occurs when a timestamp is formatted using `date`
with any timezone-related specifiers (`%Z`, `%z`) when a timezone
argument is not provided. Datetime strings already handle this case by
assuming the value is in UTC, so the fix here is to do the same for
timestamps as well.